# frontend/src/api/openapi.json
{
  "components": {
    "schemas": {
      "AccountNumber": {
        "properties": {
          "account_number": {
            "title": "Account Number",
            "type": "string"
          }
        },
        "required": [
          "account_number"
        ],
        "title": "AccountNumber",
        "type": "object"
      },
      "AccountView": {
        "properties": {
          "account_id": {
            "title": "Account Id",
            "type": "integer"
          },
          "masked_number": {
            "title": "Masked Number",
            "type": "string"
          },
          "sub_accounts": {
            "items": {
              "$ref": "#/components/schemas/SubAccountView"
            },
            "title": "Sub Accounts",
            "type": "array"
          }
        },
        "required": [
          "account_id",
          "masked_number",
          "sub_accounts"
        ],
        "title": "AccountView",
        "type": "object"
      },
      "Action": {
        "enum": [
          "approve",
          "edit",
          "reject",
          "reply_only"
        ],
        "type": "string"
      },
...
      "CandidateView": {
        "properties": {
          "fee_txn_id": {
            "title": "Fee Txn Id",
            "type": "integer"
          },
          "label": {
            "title": "Label",
            "type": "string"
          }
        },
        "required": [
          "fee_txn_id",
          "label"
        ],
        "title": "CandidateView",
        "type": "object"
      },
      "CaseStatus": {
        "enum": [
          "not_checked",
          "checking",
          "ready_to_refund",
          "recommend_no_refund",
          "needs_supervisor",
          "needs_your_call",
          "not_about_fee",
          "done"
        ],
        "type": "string"
      },
      "CaseView": {
        "properties": {
          "actions": {
            "items": {
              "$ref": "#/components/schemas/Action"
            },
            "title": "Actions",
            "type": "array"
          },
          "can_pick_fee": {
            "title": "Can Pick Fee",
            "type": "boolean"
          },
          "can_run": {
            "title": "Can Run",
            "type": "boolean"
          },
          "candidates": {
            "items": {
              "$ref": "#/components/schemas/CandidateView"
            },
            "title": "Candidates",
            "type": "array"
          },
          "checking_run_id": {
            "anyOf": [
              {
                "format": "uuid",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Checking Run Id"
          },
          "clause": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/ClauseView"
              },
              {
                "type": "null"
              }
            ]
          },
          "conversation": {
            "$ref": "#/components/schemas/ConversationView"
          },
          "decision": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/DecisionView"
              },
              {
                "type": "null"
              }
            ]
          },
          "draft": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/DraftView"
              },
              {
                "type": "null"
              }
            ]
          },
          "evidence": {
            "$ref": "#/components/schemas/EvidenceView"
          },
          "fee": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/FeeView"
              },
              {
                "type": "null"
              }
            ]
          },
          "id": {
            "title": "Id",
            "type": "integer"
          },
          "language": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Language"
          },
          "member": {
            "$ref": "#/components/schemas/MemberView"
          },
          "notes": {
            "items": {
              "$ref": "#/components/schemas/ReasonView"
            },
            "title": "Notes",
            "type": "array"
          },
          "reasons": {
            "items": {
              "$ref": "#/components/schemas/ReasonView"
            },
            "title": "Reasons",
            "type": "array"
          },
          "recommendation": {
            "$ref": "#/components/schemas/RecommendationView"
          },
          "run": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/RunView"
              },
              {
                "type": "null"
              }
            ]
          },
          "status": {
            "$ref": "#/components/schemas/CaseStatus"
          },
          "summary": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Summary"
          },
          "topic": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/Topic"
              },
              {
                "type": "null"
              }
            ]
          }
        },
        "required": [
          "id",
          "conversation",
          "member",
          "status",
          "topic",
          "language",
          "summary",
          "reasons",
          "notes",
          "recommendation",
          "fee",
          "candidates",
          "evidence",
          "clause",
          "draft",
          "run",
          "decision",
          "actions",
          "can_run",
          "can_pick_fee",
          "checking_run_id"
        ],
        "title": "CaseView",
        "type": "object"
      },
      "CheckView": {
        "properties": {
          "facts": {
            "additionalProperties": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                },
                {
                  "type": "boolean"
                },
                {
                  "items": {
                    "type": "string"
                  },
                  "type": "array"
                }
              ]
            },
            "title": "Facts",
            "type": "object"
          },
          "label": {
            "title": "Label",
            "type": "string"
          },
          "passed": {
            "title": "Passed",
            "type": "boolean"
          }
        },
        "required": [
          "label",
          "passed",
          "facts"
        ],
        "title": "CheckView",
        "type": "object"
      },
      "ClauseView": {
        "properties": {
          "doc_title": {
            "title": "Doc Title",
            "type": "string"
          },
          "section": {
            "title": "Section",
            "type": "string"
          },
          "text": {
            "title": "Text",
            "type": "string"
          }
        },
        "required": [
          "doc_title",
          "section",
          "text"
        ],
        "title": "ClauseView",
        "type": "object"
      },
      "ConversationMessage": {
        "properties": {
          "author": {
            "enum": [
              "member",
              "staff"
            ],
            "title": "Author",
            "type": "string"
          },
          "author_name": {
            "title": "Author Name",
            "type": "string"
          },
          "body": {
            "title": "Body",
            "type": "string"
          },
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          }
        },
        "required": [
          "author",
          "author_name",
          "body",
          "created_at"
        ],
        "title": "ConversationMessage",
        "type": "object"
      },
      "ConversationView": {
        "properties": {
          "messages": {
            "items": {
              "$ref": "#/components/schemas/ConversationMessage"
            },
            "title": "Messages",
            "type": "array"
          },
          "status": {
            "title": "Status",
            "type": "string"
          },
          "subject": {
            "title": "Subject",
            "type": "string"
          }
        },
        "required": [
          "subject",
          "status",
          "messages"
        ],
        "title": "ConversationView",
        "type": "object"
      },
...
      "DecisionView": {
        "properties": {
          "action": {
            "$ref": "#/components/schemas/Action"
          },
          "amount": {
            "anyOf": [
              {
                "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Amount"
          },
          "at": {
            "format": "date-time",
            "title": "At",
            "type": "string"
          },
          "by": {
            "title": "By",
            "type": "string"
          },
          "refunded": {
            "title": "Refunded",
            "type": "boolean"
          },
          "reply": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Reply"
          }
        },
        "required": [
          "by",
          "at",
          "action",
          "refunded",
          "amount",
          "reply"
        ],
        "title": "DecisionView",
        "type": "object"
      },
      "DraftView": {
        "properties": {
          "source": {
            "enum": [
              "model",
              "template"
            ],
            "title": "Source",
            "type": "string"
          },
          "text": {
            "title": "Text",
            "type": "string"
          }
        },
        "required": [
          "text",
          "source"
        ],
        "title": "DraftView",
        "type": "object"
      },
      "EvidenceView": {
        "properties": {
          "checks": {
            "items": {
              "$ref": "#/components/schemas/CheckView"
            },
            "title": "Checks",
            "type": "array"
          },
          "fee_day": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/FeeDayView"
              },
              {
                "type": "null"
              }
            ]
          },
          "refunds_in_window": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/RefundsWindowView"
              },
              {
                "type": "null"
              }
            ]
          },
          "standing": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/StandingView"
              },
              {
                "type": "null"
              }
            ]
          }
        },
        "required": [
          "fee_day",
          "refunds_in_window",
          "standing",
          "checks"
        ],
        "title": "EvidenceView",
        "type": "object"
      },
...
      "FeeDayView": {
        "properties": {
          "date": {
            "format": "date",
            "title": "Date",
            "type": "string"
          },
          "rows": {
            "items": {
              "$ref": "#/components/schemas/FeeDayRow"
            },
            "title": "Rows",
            "type": "array"
          },
          "summary": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Summary"
          }
        },
        "required": [
          "date",
          "rows",
          "summary"
        ],
        "title": "FeeDayView",
        "type": "object"
      },
      "FeeView": {
        "properties": {
          "amount": {
            "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$",
            "title": "Amount",
            "type": "string"
          },
          "date": {
            "format": "date",
            "title": "Date",
            "type": "string"
          },
          "description": {
            "title": "Description",
            "type": "string"
          },
          "fee_txn_id": {
            "title": "Fee Txn Id",
            "type": "integer"
          },
          "fee_type": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Fee Type"
          },
          "source": {
            "anyOf": [
              {
                "enum": [
                  "rule",
                  "jev",
                  "staff"
                ],
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Source"
          },
          "sub_account_name": {
            "title": "Sub Account Name",
            "type": "string"
          }
        },
        "required": [
          "fee_txn_id",
          "date",
          "amount",
          "fee_type",
          "description",
          "sub_account_name",
          "source"
        ],
        "title": "FeeView",
        "type": "object"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "MemberView": {
        "properties": {
          "accounts": {
            "items": {
              "$ref": "#/components/schemas/AccountView"
            },
            "title": "Accounts",
            "type": "array"
          },
          "name": {
            "title": "Name",
            "type": "string"
          }
        },
        "required": [
          "name",
          "accounts"
        ],
        "title": "MemberView",
        "type": "object"
      },
...
      "QueueItem": {
        "properties": {
          "amount": {
            "anyOf": [
              {
                "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Amount"
          },
          "checked_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Checked At"
          },
          "id": {
            "title": "Id",
            "type": "integer"
          },
          "member_name": {
            "title": "Member Name",
            "type": "string"
          },
          "received_at": {
            "format": "date-time",
            "title": "Received At",
            "type": "string"
          },
          "status": {
            "$ref": "#/components/schemas/CaseStatus"
          },
          "subject": {
            "title": "Subject",
            "type": "string"
          },
          "topic": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/Topic"
              },
              {
                "type": "null"
              }
            ]
          }
        },
        "required": [
          "id",
          "member_name",
          "subject",
          "received_at",
          "status",
          "topic",
          "amount",
          "checked_at"
        ],
        "title": "QueueItem",
        "type": "object"
      },
      "QueuePage": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/QueueItem"
            },
            "title": "Items",
            "type": "array"
          },
          "next_cursor": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Next Cursor"
          }
        },
        "required": [
          "items",
          "next_cursor"
        ],
        "title": "QueuePage",
        "type": "object"
      },
      "ReasonView": {
        "properties": {
          "message": {
            "title": "Message",
            "type": "string"
          },
          "next_step": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Next Step"
          }
        },
        "required": [
          "message",
          "next_step"
        ],
        "title": "ReasonView",
        "type": "object"
      },
      "RecommendationView": {
        "properties": {
          "action": {
            "enum": [
              "refund",
              "no_refund",
              "none"
            ],
            "title": "Action",
            "type": "string"
          },
          "amount": {
            "anyOf": [
              {
                "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Amount"
          }
        },
        "required": [
          "action",
          "amount"
        ],
        "title": "RecommendationView",
        "type": "object"
      },
...
      "RefundsWindowView": {
        "properties": {
          "count": {
            "title": "Count",
            "type": "integer"
          },
          "end": {
            "format": "date",
            "title": "End",
            "type": "string"
          },
          "items": {
            "items": {
              "$ref": "#/components/schemas/RefundRow"
            },
            "title": "Items",
            "type": "array"
          },
          "max_allowed": {
            "title": "Max Allowed",
            "type": "integer"
          },
          "start": {
            "format": "date",
            "title": "Start",
            "type": "string"
          }
        },
        "required": [
          "start",
          "end",
          "count",
          "max_allowed",
          "items"
        ],
        "title": "RefundsWindowView",
        "type": "object"
      },
      "RunRequest": {
        "additionalProperties": false,
        "properties": {
          "fee_txn_id": {
            "anyOf": [
              {
                "exclusiveMinimum": 0.0,
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Fee Txn Id"
          }
        },
        "title": "RunRequest",
        "type": "object"
      },
      "RunStarted": {
        "properties": {
          "run_id": {
            "format": "uuid",
            "title": "Run Id",
            "type": "string"
          }
        },
        "required": [
          "run_id"
        ],
        "title": "RunStarted",
        "type": "object"
      },
      "RunView": {
        "properties": {
          "checked_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Checked At"
          },
          "cost_usd": {
            "anyOf": [
              {
                "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Cost Usd"
          },
          "duration_ms": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Duration Ms"
          },
          "provider_mode": {
            "additionalProperties": {
              "type": "string"
            },
            "title": "Provider Mode",
            "type": "object"
          },
          "run_id": {
            "format": "uuid",
            "title": "Run Id",
            "type": "string"
          },
          "steps": {
            "items": {
              "$ref": "#/components/schemas/StepView"
            },
            "title": "Steps",
            "type": "array"
          }
        },
        "required": [
          "run_id",
          "checked_at",
          "duration_ms",
          "cost_usd",
          "provider_mode",
          "steps"
        ],
        "title": "RunView",
        "type": "object"
      },
      "StandingView": {
        "properties": {
          "below_zero": {
            "items": {
              "$ref": "#/components/schemas/BelowZero"
            },
            "title": "Below Zero",
            "type": "array"
          },
          "ok": {
            "title": "Ok",
            "type": "boolean"
          }
        },
        "required": [
          "ok",
          "below_zero"
        ],
        "title": "StandingView",
        "type": "object"
      },
      "StepView": {
        "properties": {
          "latency_ms": {
            "title": "Latency Ms",
            "type": "integer"
          },
          "node": {
            "title": "Node",
            "type": "string"
          },
          "state": {
            "enum": [
              "finished",
              "failed"
            ],
            "title": "State",
            "type": "string"
          }
        },
        "required": [
          "node",
          "state",
          "latency_ms"
        ],
        "title": "StepView",
        "type": "object"
      },
...
      "Topic": {
        "enum": [
          "fee_refund_request",
          "fee_question",
          "card_issue",
          "account_update",
          "statement_question",
          "other"
        ],
        "type": "string"
      },
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      }
    }
  },
  "info": {
    "title": "Fee Refund Agent",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
    "/cases": {
      "get": {
        "operationId": "list_cases_cases_get",
        "parameters": [
          {
            "in": "query",
            "name": "view",
            "required": false,
            "schema": {
              "default": "open",
              "enum": [
                "open",
                "done"
              ],
              "title": "View",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 50,
              "maximum": 100,
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "cursor",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "maxLength": 200,
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Cursor"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/QueuePage"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "List Cases"
      }
    },
    "/cases/{case_id}": {
      "get": {
        "operationId": "get_case_cases__case_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "case_id",
            "required": true,
            "schema": {
              "minimum": 1,
              "title": "Case Id",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CaseView"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Case"
      }
    },
    "/cases/{case_id}/accounts/{account_id}/number": {
      "get": {
        "description": "The full number, for this one request, and audited (SPEC-api; D9).",
        "operationId": "reveal_account_number_cases__case_id__accounts__account_id__number_get",
        "parameters": [
          {
            "in": "path",
            "name": "case_id",
            "required": true,
            "schema": {
              "minimum": 1,
              "title": "Case Id",
              "type": "integer"
            }
          },
          {
            "in": "path",
            "name": "account_id",
            "required": true,
            "schema": {
              "minimum": 1,
              "title": "Account Id",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AccountNumber"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Reveal Account Number"
      }
    },
...
    "/cases/{case_id}/run": {
      "post": {
        "operationId": "check_case_cases__case_id__run_post",
        "parameters": [
          {
            "in": "path",
            "name": "case_id",
            "required": true,
            "schema": {
              "minimum": 1,
              "title": "Case Id",
              "type": "integer"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "anyOf": [
                  {
                    "$ref": "#/components/schemas/RunRequest"
                  },
                  {
                    "type": "null"
                  }
                ],
                "title": "Body"
              }
            }
          }
        },
        "responses": {
          "202": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RunStarted"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Check Case"
      }
    },
...
  }
}